#23421 Move AdressHelper and ApiFactory to Service
Move these to Service namespace
